# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Wawra audio platform API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: categories
    description: Audio category operations
  - name: audios
    description: Audio content operations
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  # ── Categories ──────────────────────────────────────────
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Slug conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories/{id}:
    get:
      operationId: getCategory
      tags: [categories]
      summary: Get a single category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryBody"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /categories/slug/{slug}:
    get:
      operationId: getCategoryBySlug
      tags: [categories]
      summary: Get a category by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Audios ──────────────────────────────────────────────
  /audios:
    get:
      operationId: listAudios
      tags: [audios]
      summary: List audios (with optional search/filter)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: categoryId
          in: query
          required: false
          schema:
            type: integer
        - name: published
          in: query
          required: false
          schema:
            type: boolean
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Paginated audio list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioListResponse"

    post:
      operationId: createAudio
      tags: [audios]
      summary: Create audio entry (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAudioBody"
      responses:
        "201":
          description: Created audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audio"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /audios/featured:
    get:
      operationId: getFeaturedAudios
      tags: [audios]
      summary: Get featured/highlighted audios for homepage
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Featured audios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Audio"

  /audios/latest:
    get:
      operationId: getLatestAudios
      tags: [audios]
      summary: Get the most recently published audios
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Latest audios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Audio"

  /audios/stats:
    get:
      operationId: getAudioStats
      tags: [audios]
      summary: Get platform-wide stats for admin dashboard
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AudioStats"

  /audios/{id}:
    get:
      operationId: getAudio
      tags: [audios]
      summary: Get a single audio by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audio"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateAudio
      tags: [audios]
      summary: Update audio (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAudioBody"
      responses:
        "200":
          description: Updated audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audio"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteAudio
      tags: [audios]
      summary: Delete audio (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /audios/slug/{slug}:
    get:
      operationId: getAudioBySlug
      tags: [audios]
      summary: Get audio by slug (public)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Audio
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Audio"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Upload ──────────────────────────────────────────────
  /upload/audio:
    post:
      operationId: uploadAudioFile
      tags: [admin]
      summary: Upload an audio file and return its public URL
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /upload/image:
    post:
      operationId: uploadImageFile
      tags: [admin]
      summary: Upload a cover image and return its public URL
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Admin Auth ──────────────────────────────────────────
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session status
      responses:
        "200":
          description: Authenticated admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        audioCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - description
        - color
        - audioCount
        - createdAt
        - updatedAt

    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
      required:
        - name
        - slug

    UpdateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        color:
          type: ["string", "null"]

    Audio:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        audioUrl:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        categorySlug:
          type: ["string", "null"]
        categoryColor:
          type: ["string", "null"]
        published:
          type: boolean
        featured:
          type: boolean
        publishDate:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - slug
        - description
        - audioUrl
        - coverImageUrl
        - duration
        - categoryId
        - categoryName
        - categorySlug
        - categoryColor
        - published
        - featured
        - publishDate
        - createdAt
        - updatedAt

    AudioListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Audio"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - items
        - total
        - limit
        - offset

    CreateAudioBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        audioUrl:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        published:
          type: boolean
        featured:
          type: boolean
        publishDate:
          type: ["string", "null"]
          format: date-time
      required:
        - title
        - slug
        - audioUrl

    UpdateAudioBody:
      type: object
      properties:
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        audioUrl:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        categoryId:
          type: ["integer", "null"]
        published:
          type: boolean
        featured:
          type: boolean
        publishDate:
          type: ["string", "null"]
          format: date-time

    AudioStats:
      type: object
      properties:
        totalAudios:
          type: integer
        publishedAudios:
          type: integer
        totalCategories:
          type: integer
        featuredAudios:
          type: integer
        recentUploads:
          type: integer
      required:
        - totalAudios
        - publishedAudios
        - totalCategories
        - featuredAudios
        - recentUploads

    UploadResult:
      type: object
      properties:
        url:
          type: string
        filename:
          type: string
      required:
        - url
        - filename

    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    AdminSession:
      type: object
      properties:
        authenticated:
          type: boolean
        role:
          type: string
      required:
        - authenticated
        - role