# lib/api-zod/src/generated/types/audio.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Wawra audio platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Audio {
  id: number;
  title: string;
  slug: string;
  /** @nullable */
  description: string | null;
  audioUrl: string;
  /** @nullable */
  coverImageUrl: string | null;
  /** @nullable */
  duration: number | null;
  /** @nullable */
  categoryId: number | null;
  /** @nullable */
  categoryName: string | null;
  /** @nullable */
  categorySlug: string | null;
  /** @nullable */
  categoryColor: string | null;
  published: boolean;
  featured: boolean;
  /** @nullable */
  publishDate: Date | null;
  createdAt: Date;
  updatedAt: Date;
}